requirements.txt
Flask==2.0.1
Pillow==8.3.1
pytesseract==0.3.8
python-dotenv==0.19.0
Werkzeug==2.0.1
numpy==1.21.2
google-cloud-vision==2.4.2